+++ player/src/main/java/xyz/gianlu/librespot/player/Player.java
@@ -86,13 +86,21 @@ public class Player implements Closeable, DeviceStateHandler.Listener, PlayerSes
     // ================================ //
 
     public void volumeUp() {
+        this.volumeUp(1);
+    }
+
+    public void volumeUp(int steps) {
         if (state == null) return;
-        setVolume(Math.min(Player.VOLUME_MAX, state.getVolume() + oneVolumeStep()));
+        setVolume(Math.min(Player.VOLUME_MAX, state.getVolume() + steps * oneVolumeStep()));
     }
 
     public void volumeDown() {
+        this.volumeDown(1);
+    }
+
+    public void volumeDown(int steps) {
         if (state == null) return;
-        setVolume(Math.max(0, state.getVolume() - oneVolumeStep()));
+        setVolume(Math.max(0, state.getVolume() - steps * oneVolumeStep()));
     }
 
     private int oneVolumeStep() {
@@ -169,7 +177,7 @@ public class Player implements Closeable, DeviceStateHandler.Listener, PlayerSes
             }
 
             @Override
-            public void command(DeviceStateHandler.@NotNull Endpoint endpoint, DeviceStateHandler.@NotNull CommandBody data) throws InvalidProtocolBufferException {
+            public void command(DeviceStateHandler.@NotNull Endpoint endpoint, DeviceStateHandler.@NotNull CommandBody data) {
             }
 
             @Override
@@ -214,6 +222,8 @@ public class Player implements Closeable, DeviceStateHandler.Listener, PlayerSes
      * @param play      Whether the playback should start immediately
      */
     private void loadSession(@NotNull String sessionId, boolean play, boolean withSkip) {
+        LOGGER.debug("Loading session, id: {}, play: {}", sessionId, play);
+
         TransitionInfo trans = TransitionInfo.contextChange(state, withSkip);
 
         if (playerSession != null) {
@@ -240,6 +250,7 @@ public class Player implements Closeable, DeviceStateHandler.Listener, PlayerSes
     private void loadTrack(boolean play, @NotNull TransitionInfo trans) {
         endMetrics(playerSession.currentPlaybackId(), trans.endedReason, playerSession.currentMetrics(), trans.endedWhen);
 
+        LOGGER.debug("Loading track, id: {}, session: {}, playback: {}, play: {}", state.getCurrentPlayable(), playerSession.sessionId(), playerSession.currentPlaybackId(), play);
         String playbackId = playerSession.play(state.getCurrentPlayableOrThrow(), state.getPosition(), trans.startedReason);
         state.setPlaybackId(playbackId);
         session.eventService().sendEvent(new NewPlaybackIdEvent(state.getSessionId(), playbackId));